feat: add to train to clear data
