+++ package.json
@@ -15,12 +15,7 @@
         "start": "react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test",
-        "eject": "react-scripts eject",
-        "lint": "eslint .",
-        "lint:fix": "npm run lint -- --fix"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
+        "eject": "react-scripts eject"
     },
     "browserslist": {
         "production": [
