+++ src/components/Dialogs/Dialogs.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
+import { dialogs, messages } from '../../index';
+
 import style from './Dialogs.module.css';
 
 const DialogItem = (props) => {
@@ -16,47 +18,6 @@ const DialogItem = (props) => {
 const Message = (props) => <div className={style.message}>{props.message}</div>;
 
 const Dialogs = () => {
-    const dialogs = [
-        {
-            id: '1',
-            name: '105',
-        },
-        {
-            id: '2',
-            name: 'FrinteZz',
-        },
-        {
-            id: '3',
-            name: 'Gaming',
-        },
-        {
-            id: '4',
-            name: 'Pominki',
-        },
-        {
-            id: '5',
-            name: 'Side',
-        },
-        {
-            id: '6',
-            name: 'EFsquad',
-        },
-    ];
-
-    const messages = [
-        {
-            id: '1',
-            message: 'qq all',
-        },
-        {
-            id: '2',
-            message: 'hello world',
-        },
-        {
-            id: '3',
-            message: 'priveti4 ot Marmeladi4a',
-        },
-    ];
 
     const dialogsElements = dialogs.map((d) => <DialogItem name={d.name} id={d.id} />);
     const messagesElements = messages.map((m) => <Message message={m.message} />);
