+++ src/components/Header/Header.js
@@ -3,9 +3,9 @@ import React, { Component } from 'react';
 import style from './Header.module.css';
 
 const Header = () => (
-  <header className={style.header}>
-    <img src="" />
-  </header>
+    <header className={style.header}>
+        <img src="" />
+    </header>
 );
 
 export default Header;
