+++ src/components/Music/Music.js
@@ -3,9 +3,9 @@ import React from 'react';
 import style from './Music.module.css';
 
 const Music = (props) => (
-  <div>
-            Music
-  </div>
+    <div>
+        Music
+    </div>
 );
 
 export default Music;
