+++ src/components/Profile/MyPosts/Post/Post.js
@@ -1,15 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 import style from './Post.module.css';
 
 const Post = (props) => (
-  <div className={style.item}>
-    <img src="https://yt3.ggpht.com/a/AGF-l79zTH00bozogFe1oAjQyY7-lnxI-uZMKgonrw=s900-mo-c-c0xffffffff-rj-k-no" />
-    {props.message}
-    <div>
-      <span>like</span>
+    <div className={style.item}>
+        <img src="https://yt3.ggpht.com/a/AGF-l79zTH00bozogFe1oAjQyY7-lnxI-uZMKgonrw=s900-mo-c-c0xffffffff-rj-k-no" />
+        {props.message}
+        <div>
+            <span>like</span> {props.likesCount}
+        </div>
     </div>
-  </div>
 );
 
 export default Post;
