+++ src/components/Profile/Profile.js
@@ -1,17 +1,13 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 import MyPosts from './MyPosts/MyPosts';
 import ProfileInfo from './ProfileInfo/ProfileInfo';
 
-
-import style from './Profile.module.css';
-
-
 const Profile = () => (
-  <div>
-    <ProfileInfo />
-    <MyPosts />
-  </div>
+    <div>
+        <ProfileInfo />
+        <MyPosts />
+    </div>
 );
 
 export default Profile;
