+++ src/components/Profile/ProfileInfo/ProfileInfo.js
@@ -1,4 +1,3 @@
-/* eslint-disable jsx-a11y/alt-text */
 import React from 'react';
 
 import BACKGROUND_IMG from '../../../url_utils';
@@ -13,7 +12,7 @@ const ProfileInfo = () => {
             </div>
             <div className={style.descriptionBlock}>
                 ava+desc
-    </div>
+            </div>
         </div>
     );
 };
