+++ src/components/Settings/Settings.js
@@ -1,11 +1,9 @@
 import React from 'react';
 
-import style from './Settings.module.css';
-
 const Settings = (props) => (
-  <div>
-            Settings
-  </div>
+    <div>
+        Settings
+    </div>
 );
 
 export default Settings;
